+++ src/pytorch_lightning/loops/dataloader/evaluation_loop.py
@@ -267,10 +267,8 @@ class EvaluationLoop(DataLoaderLoop):
         self.trainer._logger_connector.reset_results()
 
     def _on_evaluation_epoch_start(self, *args: Any, **kwargs: Any) -> None:
-        """Runs ``on_epoch_start`` and ``on_{validation/test}_epoch_start`` hooks."""
+        """Runs the ``on_{validation/test}_epoch_start`` hooks."""
         self.trainer._logger_connector.on_epoch_start()
-        self.trainer._call_callback_hooks("on_epoch_start", *args, **kwargs)
-        self.trainer._call_lightning_module_hook("on_epoch_start", *args, **kwargs)
 
         hook_name = "on_test_epoch_start" if self.trainer.testing else "on_validation_epoch_start"
         self.trainer._call_callback_hooks(hook_name, *args, **kwargs)
@@ -295,8 +293,6 @@ class EvaluationLoop(DataLoaderLoop):
         self.trainer._call_callback_hooks(hook_name)
         self.trainer._call_lightning_module_hook(hook_name)
 
-        self.trainer._call_callback_hooks("on_epoch_end")
-        self.trainer._call_lightning_module_hook("on_epoch_end")
         self.trainer._logger_connector.on_epoch_end()
 
     @staticmethod