+++ tests/tests_pytorch/trainer/logging_/test_logger_connector.py
@@ -37,22 +37,14 @@ def test_fx_validator():
         "on_before_backward",
         "on_after_backward",
         "on_before_optimizer_step",
-        "on_batch_end",
-        "on_batch_start",
-        "on_before_accelerator_backend_setup",
         "on_before_zero_grad",
-        "on_epoch_end",
-        "on_epoch_start",
         "on_fit_end",
-        "on_configure_sharded_model",
         "on_fit_start",
         "on_init_end",
         "on_init_start",
         "on_exception",
         "on_load_checkpoint",
         "load_state_dict",
-        "on_pretrain_routine_end",
-        "on_pretrain_routine_start",
         "on_sanity_check_end",
         "on_sanity_check_start",
         "state_dict",
@@ -86,17 +78,13 @@ def test_fx_validator():
     }
 
     not_supported = {
-        "on_before_accelerator_backend_setup",
         "on_fit_end",
         "on_fit_start",
-        "on_configure_sharded_model",
         "on_init_end",
         "on_init_start",
         "on_exception",
         "on_load_checkpoint",
         "load_state_dict",
-        "on_pretrain_routine_end",
-        "on_pretrain_routine_start",
         "on_sanity_check_end",
         "on_sanity_check_start",
         "on_predict_batch_end",
@@ -207,14 +195,10 @@ def test_fx_validator_integration(tmpdir):
     """Tries to log inside all `LightningModule` and `Callback` hooks to check any expected errors."""
     not_supported = {
         None: "`self.trainer` reference is not registered",
-        "on_before_accelerator_backend_setup": "You can't",
         "setup": "You can't",
         "configure_sharded_model": "You can't",
-        "on_configure_sharded_model": "You can't",
         "configure_optimizers": "You can't",
         "on_fit_start": "You can't",
-        "on_pretrain_routine_start": "You can't",
-        "on_pretrain_routine_end": "You can't",
         "train_dataloader": "You can't",
         "val_dataloader": "You can't",
         "on_before_batch_transfer": "You can't",
@@ -256,7 +240,6 @@ def test_fx_validator_integration(tmpdir):
     not_supported.update(
         {
             # `lightning_module` ref is now present from the `fit` call
-            "on_before_accelerator_backend_setup": "You can't",
             "test_dataloader": "You can't",
             "on_test_model_eval": "You can't",
             "on_test_model_train": "You can't",